Profile with Sweetalert2

+++ frontend/src/screens/Profile.jsx
@@ -6,6 +6,7 @@ import ProfileHeader from "../components/profile/ProfileHeader";
 import axios from "axios";
 import { Products_Mockeados } from "../constants/const";
 import { useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const Profile = () => {
   const redirect = useNavigate();
@@ -34,12 +35,52 @@ const Profile = () => {
   };
 
   const handleDelete = async (id) => {
-    try {
-      await axios.delete(`http://localhost:3000/api/products/${id}`);
-      const updatedProducts = products.filter((product) => product.id !== id);
-      setProducts(updatedProducts);
-    } catch (error) {
-      console.error("Error eliminando producto:", error);
+    const response = await Swal.fire({
+      title: 'Advertencia',
+      text: '¿Está seguro que quiere eliminar el producto?',
+      icon: 'question',
+      showDenyButton: true,
+      denyButtonText: "No",
+      confirmButtonText: 'Sí',
+      confirmButtonColor: 'green',
+    });
+  
+    if (response.isConfirmed) {
+      try {
+        await axios.delete(`http://localhost:3000/api/products/${id}`);
+        const updatedProducts = products.filter((product) => product.id !== id);
+        setProducts(updatedProducts);
+        Swal.fire({
+          title: 'Éxito',
+          text: 'El producto se eliminó correctamente',
+          icon: 'success',
+          confirmButtonColor: 'green',
+        });
+      } catch (error) {
+        console.error("Error eliminando producto:", error);
+        Swal.fire({
+          title: 'Error',
+          text: 'Hubo un problema al eliminar el producto.',
+          icon: 'error',
+          confirmButtonColor: 'green',
+        });
+      }
+    } 
+    else if (response.isDenied) {
+      Swal.fire({
+        title: 'Información',
+        text: 'El producto no se eliminó.',
+        icon: 'info',
+        confirmButtonColor: 'green',
+      });
+    } 
+    else {
+      Swal.fire({
+        title: 'Error',
+        text: 'Ocurrió un error inesperado.',
+        icon: 'error',
+        confirmButtonColor: 'green',
+      });
     }
   };
 
@@ -72,6 +113,13 @@ const Profile = () => {
     };
     
     Products_Mockeados.push(newProduct);
+    Swal.fire({
+      title: 'Operación exitosa!',
+      text: 'Producto publicado',
+      icon: 'success',
+      confirmButtonText: 'OK',
+      confirmButtonColor: '#28a745'
+    })
     redirect("/home");
   };
